+++ src/completions/seventyEightHundredBasicCompletion.ts
@@ -302,6 +302,28 @@ export class SeventyEightHundredBasicCompletion extends CompletionBase {
             },
             ' ');
 
+        // drawhiscores
+        let drawhiscoresProvider = vscode.languages.registerCompletionItemProvider(
+            this.Id,
+            {
+                // return list of available language methods
+                provideCompletionItems(document: vscode.TextDocument, position: vscode.Position) {
+                    // get all text until the `position` and check if it reads `return `
+                    let linePrefix = document.lineAt(position).text.substr(0, position.character);
+                    if (!linePrefix.endsWith('drawhiscores ')) {
+                        return undefined;
+                    }
+
+                    return [
+                        new vscode.CompletionItem('attract', vscode.CompletionItemKind.Keyword),
+                        new vscode.CompletionItem('single', vscode.CompletionItemKind.Keyword),
+                        new vscode.CompletionItem('player1', vscode.CompletionItemKind.Keyword),
+                        new vscode.CompletionItem('player2', vscode.CompletionItemKind.Keyword)
+                    ];
+                }
+            },
+            ' ');
+
         // tallsprite
         let tallsrpiteModeProvider = vscode.languages.registerCompletionItemProvider(
             this.Id,
@@ -373,7 +395,7 @@ export class SeventyEightHundredBasicCompletion extends CompletionBase {
             ' ');
 
         // add items           
-        context.subscriptions.push(ifProvider, returnProvider, romsizeProvider, displayModeProvider, shakescreenModeProvider, tallsrpiteModeProvider, setProvider);
+        context.subscriptions.push(ifProvider, returnProvider, romsizeProvider, displayModeProvider, shakescreenModeProvider, drawhiscoresProvider, tallsrpiteModeProvider, setProvider);
     }
 
 }